Initial commit from 1 to 1 tactics 1.6_release - 8 cans to collect enabled - ultra enabled - ultra sequence fixed - .....

+++ Pomidor15/src/settings.c
@@ -52,6 +52,6 @@ double settingDistanceToRunFromEnemy = 25;
  */
 int settingTimeBlockedToFail = 600;
 
-double settingVelocityToBlocked = 30;
+int settingVelocityToBlocked = 30;
 
-double settingPWMToBlocked = 0.25;
+int settingPWMToBlocked = 0.25;
